migrate to pypdf from PyPDF2 due to security issue of the library
