+++ requirements.txt
@@ -1,1 +1,2 @@
-PyPDF2==3.0.1
+pypdf==5.4.0
+typing_extensions==4.13.2
