+++ src/crop_page.py
@@ -1,8 +1,8 @@
 import sys
-import PyPDF2 as pdf
+import pypdf as pdf
 
 from typing import List
-from PyPDF2.generic import RectangleObject
+from pypdf.generic import RectangleObject
 
 
 def crop_page(
@@ -11,11 +11,10 @@ def crop_page(
     crop_bbox: List[float],
 ):
     # read pdf file
-    with open(input_file, 'rb') as fp:
-        pdf_reader = pdf.PdfFileReader(fp)
+    pdf_reader = pdf.PdfReader(input_file)
 
     # create a writer for black pdf
-    pdf_writer = pdf.PdfFileWriter()
+    pdf_writer = pdf.PdfWriter()
 
     # get page by (page index - 1)
     page = pdf_reader.pages[page_number - 1]
@@ -23,8 +22,7 @@ def crop_page(
     height = page.mediabox.height
 
     # crop by updating cropbox property of the page
-    # Rectangle Object: [x1, y1, x2, y2] => (left, bottom, right, top)
-    # crops 25% from the top and 20% from the right
+    # Rectangle Object for cropping edges => (left, bottom, right, top)
     top, left, bottom, right = crop_bbox
     page.cropbox = RectangleObject((width * left, height * bottom, width * top, height * right))
 
@@ -46,8 +44,8 @@ if __name__ == '__main__':
     crop_bbox = crop_bbox.split(',')
     crop_bbox = [float(p) for p in crop_bbox]
 
-    # crop bbox is assumed to be: [top, left, bottom, right]; 
-    # e.g: [0.2, 0.0, 0.0, 0.25] => crops 20% from top, 25% from right side of a page.
+    # crop bbox is assumed to be [top_fraction, left_fraction, bottom_fraction, right_fraction] for cropping sizes; 
+    # e.g: [0.75, 0.0, 0.0, 0.8] => crops 25% from top, 20% from right side of a page.
     crop_page(
         input_file=input_file,
         page_number=page_number,
