+++ src/extract_pages.py
@@ -1,5 +1,5 @@
 import sys
-import PyPDF2 as pdf
+import pypdf as pdf
 
 
 def extract_pages(
@@ -8,30 +8,26 @@ def extract_pages(
     end_page: int,
 ):
     # create reader for input pdf file
-    with open(input_file, 'rb') as fp:
-        inputpdf = pdf.PdfFileReader(fp)
+    pdf_reader = pdf.PdfReader(input_file)
 
     # create output pdf
-    outputpdf = pdf.PdfFileWriter()
+    pdf_writer = pdf.PdfWriter()
 
     # page count starts from 0 index
-    if start_page <= 0:
-        start_page = 0
-    else:
-        start_page -= 1
+    start_page = 0 if start_page < 0 else start_page - 1
     
-    if end_page >= len(inputpdf.pages):
-        end_page = len(inputpdf.pages) - 1
+    if end_page > len(pdf_reader.pages):
+        end_page = len(pdf_reader.pages)
     
     # Since range is used to iterate, end_page doesn't need to be decremented by one.
     # Because range doesn't include end index in iteration.
     for i in range(start_page, end_page, 1):
-        outputpdf.add_page(inputpdf.pages[i])
+        pdf_writer.add_page(pdf_reader.pages[i])
 
     # save output pdf into file
-    output_file = input_file.replace('.pdf', f'_{start_page}-{end_page}_pages.pdf')
+    output_file = input_file.replace('.pdf', f'_{start_page+1}-{end_page}_pages.pdf')
     with open(output_file, 'wb') as fp:
-        outputpdf.write(fp)
+        pdf_writer.write(fp)
 
 
 if __name__ == '__main__':
